requirements.txt
# Core dependencies
langchain>=0.1.0
langchain_community>=0.0.10
pydantic>=2.5.0
pydantic-ai>=0.0.24
python-dotenv>=1.0.0
pypdf>=3.17.0

# LLM and Tracing
openai>=1.0.0  # For OpenRouter API compatibility
langwatch>=0.1.0   # LangWatch tracing
aiohttp>=3.9.0     # Async HTTP client for API calls

# Document Processing
python-docx>=0.8.11
pdfplumber>=0.10.0
pandas>=2.0.0
numpy>=1.24.0

# API and Web
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.3.0
isort>=5.12.0
mypy>=1.4.0
flake8>=6.1.0
